+++ lib/errors/index.ts
@@ -1,7 +1,5 @@
 export { AlreadyStoppedError } from './AlreadyStoppedError';
 export { DiagnosticError } from './DiagnosticError';
-export { InvalidOptionError } from './InvalidOptionError';
 export { InvalidOptionsError } from './InvalidOptionsError';
 export { InvalidStateError } from './InvalidStateError';
-export { PromiseTimedOutError } from './PromiseTimedOutError';
 export { UnsupportedError } from './UnsupportedError';