+++ CHANGELOG.md
@@ -21,6 +21,26 @@ Changes
 
 * `MediaConnectionBitrateTest` no longer perform any analysis on bitrate data and `MediaConnectionBitrateTest.Report.didPass` is no longer available. Your application will need to analyze the bitrate values in `MediaConnectionBitrateTest.Report.values` and `MediaConnectionBitrateTest.Report.averageBitrate` properties to determine whether or not the values are acceptable.
 
+* `DiagnosticError` can now be imported directly. Example:
+
+  ```ts
+  import {
+    DiagnosticError,
+    MediaConnectionBitrateTest,
+    testMediaConnectionBitrate,
+  } from '@twilio/rtc-diagnostics';
+
+  const test = testMediaConnectionBitrate(...);
+
+  test.on(MediaConnectionBitrateTest.Events.Error, (error: DiagnosticError) => {
+    console.log(error);
+  });
+  ```
+
+* Updated test names to `audio-input-test`, `audio-output-test`, and `media-connection-bitrate-test` for consistency.
+
+* Removed unused `promise-timed-out` and `invalid-option` error.
+
 1.0.0-beta1 (July 29, 2020)
 ============================